move inspect to the canister module
